# ShopApi/nginx/data/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "s.x-01.ru API",
    "description": "ASP.NET Core Web API for mebel cluster CMS",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/igoradigey01/webapi"
    },
    "license": {
      "name": "MIT License"
    },
    "version": "v2"
  },
  "paths": {
...
    "/api/Order/GetItem/{id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "GetItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "orderNo": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "nullable": true
          },
          "ownerPhone": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "orderAdress": {
            "type": "string",
            "nullable": true
          },
          "orderPickup": {
            "type": "boolean"
          },
          "orderNote": {
            "type": "string",
            "nullable": true
          },
          "customerFullName": {
            "type": "string",
            "nullable": true
          },
          "customerId": {
            "type": "string",
            "nullable": true
          },
          "customerPhone": {
            "type": "string",
            "nullable": true
          },
          "customerMail": {
            "type": "string",
            "nullable": true
          },
          "payment_total": {
            "type": "number",
            "format": "float"
          },
          "total": {
            "type": "number",
            "format": "float"
          },
          "paymentStateId": {
            "type": "integer",
            "format": "int32"
          },
          "orderStateId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}